Extract save parsing into separate API
Save parsing is too resource intense for edge compute, so by extracting
save parsing into a standalone service (that is still serverless
compatible), the remaining server-side logic can be housed elsewhere.
