+++ Cargo.toml
@@ -1,11 +1,13 @@
 [workspace]
 members = [
+    "src/api",
     "src/applib",
-    "src/applib-node",
     "src/cli",
     "src/eu4game",
+    "src/eu4game-data",
     "src/schemas",
     "src/vic3save",
+    "src/wasm-app",
     "src/wasm-compress",
     "src/wasm-ck3",
     "src/wasm-eu4",
