+++ dev/api.dockerfile
@@ -1,0 +1,3 @@
+FROM gcr.io/distroless/static-debian11
+COPY ./pdx-tools-api /app
+CMD ["/app"]
