+++ src/app/.env.development
@@ -9,3 +9,5 @@ S3_SECRET_KEY=minio12minio
 S3_PORT=9000
 S3_ENDPOINT=http://localhost:$S3_PORT
 SESSION_SECRET="fdviwdAbBi2DWRh0fN4sBGEdVngfRhRj"
+PARSE_API_PORT=8080
+PARSE_API_ENDPOINT=http://localhost:$PARSE_API_PORT
