+++ src/app/src/lib/fetch.ts
@@ -8,6 +8,8 @@ export async function fetchOk(...args: Parameters<typeof fetch>) {
   return resp;
 }
 
-export async function fetchOkJson(...args: Parameters<typeof fetch>) {
+export async function fetchOkJson<T = any>(
+  ...args: Parameters<typeof fetch>
+): Promise<T> {
   return fetchOk(...args).then((x) => x.json());
 }
