+++ src/app/src/server-lib/save-parser.ts
@@ -1,0 +1,13 @@
+import { fetchOkJson } from "@/lib/fetch";
+import { check } from "@/lib/isPresent";
+import { ParseResult } from "./save-parsing-types";
+export type * from "./save-parsing-types";
+
+export const parseSave = (data: BodyInit) =>
+  fetchOkJson<ParseResult>(check(process.env["PARSE_API_ENDPOINT"]), {
+    method: "POST",
+    body: data,
+    headers: {
+      "Content-Type": "application/octet-stream",
+    },
+  });
