--- src/applib-node/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "applib-node"
-version = "0.1.0"
-authors = ["nickbabcock <nbabcock19@hotmail.com>"]
-edition = "2021"
-publish = false
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-napi = "2"
-napi-derive = "2"
-applib = { path = "../applib" }
-eu4game = { path = "../eu4game" }
-serde_json = "1"
-serde = {version = "1", features = ["derive"]}
-
-[build-dependencies]
-napi-build = "2"
