--- src/applib-node/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    napi_build::setup();
-}
