+++ src/applib/src/bin/types.rs
@@ -1,0 +1,7 @@
+// Import types for specta to export
+#[allow(unused_imports)]
+use applib::*;
+
+fn main() {
+    specta::export::ts("src/app/src/server-lib/save-parsing-types.ts").unwrap();
+}
