+++ src/eu4game-data/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "eu4game-data"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+embedded = []
+tsify = ["dep:tsify", "dep:wasm-bindgen"]
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+tsify = { version = "0.4.5", default-features = false, optional = true }
+wasm-bindgen = { version = "0.2", default-features = false, optional = true }
+
+[build-dependencies]
+regex = "1.5"
