+++ src/eu4game/Cargo.toml
@@ -9,10 +9,11 @@ publish = false
 default = ["embedded", "miniz"]
 miniz = ["eu4save/miniz"]
 libdeflate = ["eu4save/libdeflate"]
-embedded = []
+embedded = ["eu4game-data/embedded"]
 
 [dependencies]
 jomini = { version = "0.22.0" }
+eu4game-data = { path = "../eu4game-data" }
 eu4save = { git = "https://github.com/rakaly/eu4save.git", default-features = false }
 schemas = { path = "../schemas", features = ["inline"] }
 serde = { version = "1", features = ["derive"] }
