+++ src/wasm-app/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "wasm-app"
+version = "0.1.0"
+authors = ["Nick Babcock <nbabcock19@hotmail.com>"]
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+wasm-bindgen = { version = "0.2" }
+tsify = { version = "0.4.5", default-features = false, features = ["js"] } 
+serde = { version = "1.0", features = ["derive"] }
+eu4game-data = { path = "../eu4game-data", features = ["tsify"] }
+jomini = { version = "0.22.0" }
+
+[package.metadata.wasm-pack.profile.release]
+wasm-opt = false
